database commit

+++ packetTracerBase/src/main/java/com/example/packettracerbase/model/Location.java
@@ -6,6 +6,6 @@
     @Embeddable
     @Data
     public class Location {
-        protected Float latitude;
-        protected Float longitude;
+        protected Double latitude;
+        protected Double longitude;
     }
